Added reticle and fixed mouse aim issue
For the reticle, I am using one of the crosshairs that comes with the free Kenney Crosshair Pack https://kenney.nl/assets/crosshair-pack. Since adding the cinemachine camera, there has been an ongoing issue where the mouse would not be locked to the game, meaning the cursor would move away from where the player was aiming. This has been fixed simply by setting the cursor lock state to "Locked" when in game, and setting it to "Confined" when accessing UI.