Imported combo package, created basic spawner,. and fixed player aim issues
From the previous game I worked on "Hellknigh", I have taken the relevant assets used in the games "combo" system such as the scriptable objects, interfaces, and scripts. This is imported as a package and can be reused. I have also created a basic spawner that spawns enemies at certain intervals, with the number of enemies scaling over time (will need to have a proper look at values for balancing). Lastly I fixed an issue where the player's projectiles were not being fired close enough to the direction of the crosshair.