Nearly fully implemented stat upgrade system, most functionality is there
Made a start on the stat upgrade system and so far we have a StatManagerComponent which contains the functionality to alter given stats based on their increaseValue set in their StatObject class instance. The ApplyStatIncrease() function is the main public function used and will be used in a way so that anytime the players relevant stats would be accessed, there is an option to either get the base or modified version of the stat.