Fixed GetStat errors and restructuredfolders
There were a few errors revolving around the GetStat() functions. One was due to the name of the MaxHealth enum being Health instead, so when other components refered to it, it was not found. Another bug was caused by the players StatManager component being null at the time it was being accessed by the projectile scripts. To fix this I simply changed all the GetStat functions to check that StatManager is not null before getting the modified stat, if it is null (which should only be on the spawn of some objects) we just use the base stat instead.