Create sniper prefabs and fixed issue with multiple level ups
Started on implementing sniper, just need to create a unique sniper bullet script that works with the StatusEffectComponent I need to work on. Also fixed a possible issue where level up menus would overlap and because the mouse would be locked after selecting a gun option, the player wouldnt be able to select the rest of the UI's. Fixed this mainly in HUD script.